Rename kivydesignerapp Let's try to keep module names simpler, to a max of two words if possible
