+++ kivydesigner/KivyDesigner.kv
@@ -1,4 +1,4 @@
-#:import ICON_PATH kivydesigner.kivydesignerapp.ICON_PATH
+#:import ICON_PATH kivydesigner.kivydesigner.ICON_PATH
 #:set menubar_color (0.2, 0.2, 0.2, 1)
 
 RootWidget:
